Revert "Started working on dimensions"
This reverts commit 0e3bb9555ff825077e5bc9bc33aa52b881423c1c.
