+++ build.gradle
@@ -47,7 +47,6 @@ dependencies {
 
 	modCompile "net.fabricmc:fabric-loader:${project.loader_version}"
 	modCompile "net.fabricmc.fabric-api:fabric-api:${project.fabric_version}"
-	modCompile "net.fabricmc.fabric-api:fabric-dimensions-v1:0.2.0+969462033c"
 
 	modImplementation "artificemc:artifice:0.3.5"
 	include "artificemc:artifice:0.3.5"
