--- src/main/java/io/github/vampirestudios/raa/api/namegeneration/biomes/Norwegian.java
@@ -1,19 +1,0 @@
-package io.github.vampirestudios.raa.api.namegeneration.biomes;
-
-import io.github.vampirestudios.raa.api.namegeneration.INameGenerator;
-
-import java.util.Map;
-
-public class Norwegian implements INameGenerator {
-
-    @Override
-    public String generate() {
-        return null;
-    }
-
-    @Override
-    public Map<String, String> getSpecialCharatersMap() {
-        return null;
-    }
-
-}
