+++ src/main/java/io/github/vampirestudios/raa/blocks/LayeredOreBlock.java
@@ -1,6 +1,7 @@
 package io.github.vampirestudios.raa.blocks;
 
-import io.github.vampirestudios.raa.generation.materials.Material;
+import io.github.vampirestudios.raa.api.enums.GeneratesIn;
+import io.github.vampirestudios.raa.materials.Material;
 import io.github.vampirestudios.raa.RandomlyAddingAnything;
 import io.github.vampirestudios.raa.api.enums.OreTypes;
 import io.github.vampirestudios.raa.utils.Rands;
