--- src/main/java/io/github/vampirestudios/raa/generation/dimensions/CustomDimensionType.java
@@ -1,4 +1,0 @@
-package io.github.vampirestudios.raa.generation.dimensions;
-
-public class CustomDimensionType {
-}
