+++ src/main/java/io/github/vampirestudios/raa/items/RAAAxeItem.java
@@ -1,6 +1,6 @@
 package io.github.vampirestudios.raa.items;
 
-import io.github.vampirestudios.raa.generation.materials.Material;
+import io.github.vampirestudios.raa.materials.Material;
 import net.minecraft.item.AxeItem;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ToolMaterial;
