+++ src/main/java/io/github/vampirestudios/raa/materials/CustomArmorMaterial.java
@@ -1,4 +1,4 @@
-package io.github.vampirestudios.raa.generation.materials;
+package io.github.vampirestudios.raa.materials;
 
 import io.github.vampirestudios.raa.registries.Materials;
 import io.github.vampirestudios.raa.api.enums.OreTypes;
