--- src/main/java/io/github/vampirestudios/raa/registries/Biomes.java
@@ -1,4 +1,0 @@
-package io.github.vampirestudios.raa.registries;
-
-public class Biomes {
-}
