--- src/main/java/io/github/vampirestudios/raa/registries/Mobs.java
@@ -1,7 +1,0 @@
-package io.github.vampirestudios.raa.registries;
-
-public class Mobs {
-
-
-
-}
